# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SOS Sales API",
    "version": "1.0.0",
    "description": "API do SOS Sales — Sistema Operacional Comercial (TX Commercial Core). Todas as rotas exigem Bearer JWT emitido pelo Supabase Auth."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/workspaces/{workspaceId}/journeys/{journeyId}/outcomes": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "workspaceId",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "journeyId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/v1/billing/abacatepay/charges": {
      "post": {
        "tags": [
          "Billing"
        ],
        "description": "Cria uma cobrança PIX via AbacatePay com link de pagamento e chave copia-e-cola.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "workspaceId",
                  "customerName",
                  "customerPhone",
                  "customerEmail",
                  "productName",
                  "priceInCents"
                ],
                "properties": {
                  "workspaceId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "customerName": {
                    "type": "string"
                  },
                  "customerPhone": {
                    "type": "string"
                  },
                  "customerEmail": {
                    "type": "string"
                  },
                  "customerTaxId": {
                    "type": "string"
                  },
                  "productName": {
                    "type": "string"
                  },
                  "priceInCents": {
                    "type": "number",
                    "minimum": 100
                  },
                  "externalId": {
                    "type": "string"
                  },
                  "metadata": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "billingId": {
                      "type": "string"
                    },
                    "url": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "amount": {
                      "type": "number"
                    },
                    "pixCopiaECola": {
                      "type": "string"
                    },
                    "pixQrCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/webhooks/abacatepay": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "description": "Webhook público para receber confirmação de pagamentos do AbacatePay.",
        "parameters": [
          {
            "name": "webhookSecret",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16
            }
          },
          {
            "name": "X-Webhook-Signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "received": {
                      "type": "boolean"
                    },
                    "event": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:4334",
      "description": "Local Dev"
    },
    {
      "url": "https://sos.mct.com.br",
      "description": "Produção VPS"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ]
}